Prevent holes in data in a buffer chain.
When starting a nested message, 4 bytes are reserved for the length. If these 4
bytes span a buffer boundary, the previous implementation would skip the
remaining bytes in the current buffer and place all 4 reserved bytes at the
beginning of the next buffer. These gaps and the end of internal buffers in a
chain are problematic for our use case. This change allows the reserved bytes
to span 2 buffers.
- Replace the uint8_t* size_field_ in ScatteredBufferWriter with a ReservedBytes
  class, which is a wrapper around 2 pointers and a length available in the
  first pointer.
- Update users of the size_field_ to use the new interface.